requirements.txt
# 战双角色识别系统依赖

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# Web框架
flask>=3.0.0

# 图像处理
pillow>=10.0.0
opencv-python>=4.0.0

# 数据处理
numpy<2.0.0
matplotlib>=3.7.0

# 可选依赖（用于数据收集和标注）
# selenium>=4.15.0
# requests>=2.31.0
# labelImg>=1.8.6
